Program cs is added

+++ Core/Entities/Druggist.cs
@@ -13,7 +13,9 @@ namespace Core.Entities
         public string Name { get; set; }
         public string Surname { get; set; }
         public byte Age { get; set; }
-        public byte Experience { get; set; }
+        public double Experience { get; set; }
         public DrugStore DrugStore { get; set; }
+
+
     }
 }
